refactor(board): move BoardButton accessors out of types.ts
getSpokenText and getNavigationTargetId are runtime derivations over a
BoardButton, not type declarations. Move them to a new feature-root
board-button.ts so types.ts holds only types, and so the accessors sit
in one place reachable by all consumers (board-viewer, activation,
message) without any subfolder owning logic the others depend on.
